Include Git submodules in deploy, if detected.
If a .gitmodules file is detected in the repository the tar file will be built using git-archive-all executable, if found.  If git-archive-all is not installed, a warning will be issued and the tar file will be built as it was previously. git-archive-all is available at https://github.com/Kentzo/git-archive-all and can be installed on Mac OS X using homebrew.